+++ src/Blocks/Composite.jl
@@ -20,13 +20,6 @@ abstract supertype which composite blocks will inherit from.
 """
 abstract type CompositeBlock{N, T} <: MatrixBlock{N, T} end
 
-"""
-    blocks(composite_block)
-
-get an iterator that iterate through all sub-blocks.
-"""
-function blocks end
-
 const GeneralComposite = Union{CompositeBlock, Sequential}
 
 function map!(f::Function, dst::GeneralComposite, itr)
@@ -39,20 +32,7 @@ function map!(f::Function, dst::GeneralComposite, itr)
 end
 
 function parameter_type(c::GeneralComposite)
-    promote_type([parameter_type(each) for each in blocks(c)]...)
-end
-
-function nparameters(c::GeneralComposite)
-    count = 0
-    for each in blocks(c)
-        count += nparameters(each)
-    end
-    count
-end
-
-function parameters(c::CompositeBlock)
-    itr = Iterators.filter(x->isprimitive(x) && hasparameter(x), traverse(c))
-    Iterators.flatten(parameters(each) for each in itr)
+    promote_type([parameter_type(each) for each in subblocks(c)]...)
 end
 
 #################
@@ -72,7 +52,7 @@ each sub-block's number of parameters.
 function dispatch!(c::CompositeBlock, itr)
     @assert nparameters(c) == length(itr) "number of parameters does not match"
 
-    blkitr = Iterators.filter(x->isprimitive(x) && hasparameter(x), traverse(c))
+    blkitr = Iterators.filter(x->x isa PrimitiveBlock && hasparameter(x), traverse(c))
     pitr = itr
     for each in blkitr
         dispatch!(each, Iterators.take(pitr, nparameters(each)))
@@ -84,7 +64,7 @@ end
 function dispatch!(f::Function, c::CompositeBlock, itr)
     @assert nparameters(c) == length(itr) "number of parameters does not match"
 
-    blkitr = Iterators.filter(x->isprimitive(x) && hasparameter(x), traverse(c))
+    blkitr = Iterators.filter(x->x isa PrimitiveBlock && hasparameter(x), traverse(c))
     pitr = itr
     for each in blkitr
         dispatch!(f, each, Iterators.take(pitr, nparameters(each)))