+++ src/Blocks/Core.jl
@@ -9,6 +9,14 @@ qubits.
 abstract type AbstractBlock end
 
 """
+    subblocks(blk::AbstractBlock) -> Tuple
+
+return a tuple of all sub-blocks in this block.
+"""
+function subblocks end
+subblocks(blk::AbstractBlock) = ()
+
+"""
     apply!(reg, block, [signal])
 
 apply a `block` to a register `reg` with or without a cache signal.
@@ -37,26 +45,34 @@ Test whether this operator is reflexive.
 function isreflexive end
 
 """
-    nparameters(x) -> Integer
+    niparameters(x) -> Integer
 
 Returns the number of parameters of `x`.
 """
-function nparameters end
-nparameters(::Type{X}) where {X <: AbstractBlock} = 0
-nparameters(x::AbstractBlock) = length(parameters(x))
+function npiarameters end
+niparameters(::Type{X}) where {X <: AbstractBlock} = 0
+niparameters(x::AbstractBlock) = length(iparameters(x))
 
 """
-    parameters(block) -> Vector
+    iparameters(block) -> Vector
 
-Returns a list of all parameters in block.
+Returns a list of all intrinsic (not from sublocks) parameters in block.
 """
-function parameters end
-parameters(x::AbstractBlock) = ()
+function iparameters end
+iparameters(x::AbstractBlock) = ()
 
 """
-    parameters(block) -> Type
+    setparameters!([elementwisefunction], r::AbstractBlock, params) -> AbstractBlock
 
-the type of parameters.
+set intrinsics parameter for block.
+"""
+function setiparameters end
+setiparameters!(r::AbstractBlock, params) = r
+
+"""
+    parameter_type(block) -> Type
+
+the type of iparameters.
 """
 function parameter_type end
 parameter_type(x::AbstractBlock) = Bool
@@ -72,7 +88,7 @@ function mat end
     dispatch!(block, params)
     dispatch!(block, params...)
 
-dispatch parameters to this block.
+dispatch (using pop!) parameters to this block.
 """
 function dispatch! end
 
@@ -82,13 +98,30 @@ function dispatch! end
 define the style to print this block
 """
 function print_block(io::IO, block)
-
-@static if VERSION < v"0.7-"
-    print(io, summary(block))
-else
     summary(io, block)
 end
 
+"""
+    nparameters(c::AbstractBlock) -> Int
+
+number of parameters, including parameters in sublocks.
+"""
+function nparameters(c::AbstractBlock)
+    count = 0
+    for each in subblocks(c)
+        count += nparameters(each)
+    end
+    count
 end
 
-isprimitive(blk::AbstractBlock) = false
+"""
+    parameters(c::AbstractBlock, output=Float64[]) -> Vector
+
+get all parameters including sublocks.
+"""
+function parameters(c::AbstractBlock, output=Float64[])
+    for blk in subblocks(c)
+        append!(output, iparameters(blk))
+    end
+    output
+end