+++ src/Blocks/Roller.jl
@@ -28,7 +28,7 @@ end
 
 copy(m::Roller) = typeof(m)(m.blocks)
 
-blocks(m::Roller) = m.blocks
+subblocks(m::Roller) = m.blocks
 addrs(m::Roller) = cumsum([[1]; [nqubits(b) for b in m.blocks[1:end-1]]])
 isunitary(m::Roller) = all(isunitary, m.blocks)
 ishermitian(m::Roller) = all(ishermitian, m.blocks)