+++ src/Interfaces/DeprecationWarnings.jl
@@ -1,0 +1,6 @@
+using Base.depwarn
+
+function blocks(blk::AbstractBlock)
+    depwarn("`blocks` will be renamed to `subblocks` to avoid confusion.")
+    subblocks(blk)
+end